scalefs: Fix crash while looking up mnode_to_inode mapping
Once in a while, under heavy stress, the LFS Smallfile benchmark used to crash
in the following way:

cpu0-lfs-smallfile: panic: unlink_old_inode: Mapping for mnode# 431292417 does not exist

The stack trace pointed to this call sequence: unlink_old_inode() -> get_inode().
Thus, we were trying to unlink an mnode from its parent directory, but the
parent directory's mnode didn't really have a corresponding inode.

It finally turned out that the mfs_unlink_operation uses u32 as the data type to
represent the parent mnode, rather than the full u64. Thus, the mnode number of
the parent directory used to get truncated and we would thus perform an invalid
lookup on the mnode_to_inode table. For an error as small as this, it consumed
an exasperating amount of debugging time and effort! :-(

+++ include/scalefs.hh
@@ -746,7 +746,7 @@ class mfs_operation_unlink: public mfs_operation
   public:
     NEW_DELETE_OPS(mfs_operation_unlink);
 
-    mfs_operation_unlink(mfs_interface *p, u64 t, u64 mn, u32 pt, char nm[])
+    mfs_operation_unlink(mfs_interface *p, u64 t, u64 mn, u64 pt, char nm[])
       : mfs_operation(p, t), mnode(mn), parent(pt)
     {
       name = new char[DIRSIZ];
